+++ src/router/segit.tsx
@@ -1,0 +1,8 @@
+import { RouteObject } from "react-router-dom";
+import Home from "../pages/segit_pages";
+
+const segitRoutes: RouteObject[] = [
+  { path: "/segit", element: <Home /> },
+];
+
+export default segitRoutes;